+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/skaffold/SyncMapTask.java
@@ -78,7 +78,10 @@ public class SyncMapTask extends DefaultTask {
 
       try {
         String syncMapJson =
-            PluginConfigurationProcessor.getSkaffoldSyncMap(configuration, projectProperties);
+            PluginConfigurationProcessor.getSkaffoldSyncMap(
+                configuration,
+                projectProperties,
+                jibExtension.getSkaffold().getSync().getExcludes());
 
         System.out.println();
         System.out.println("BEGIN JIB JSON: SYNCMAP/1");