+++ jib-gradle-plugin/src/test/java/com/google/cloud/tools/jib/gradle/JibExtensionTest.java
@@ -20,6 +20,8 @@ import com.google.cloud.tools.jib.api.ImageFormat;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
+import java.io.File;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.Collections;
@@ -238,6 +240,35 @@ public class JibExtensionTest {
     Assert.assertEquals(
         fakeProject.getProjectDir().toPath().resolve(Paths.get("path/to/tar")),
         testJibExtension.getOutputPaths().getTarPath());
+  }
+
+  @Test
+  public void testSkaffold() {
+    testJibExtension.skaffold(
+        skaffold -> {
+          skaffold.sync(sync -> sync.setExcludes(fakeProject.files("sync1", "sync2")));
+          skaffold.watch(
+              watch -> {
+                watch.setBuildIncludes(ImmutableList.of("watch1", "watch2"));
+                watch.setIncludes("watch3");
+                watch.setExcludes(ImmutableList.of(new File("watch4")));
+              });
+        });
+    Path root = fakeProject.getRootDir().toPath();
+    Assert.assertEquals(
+        ImmutableSet.of(
+            root.resolve("sync1").toAbsolutePath(), root.resolve("sync2").toAbsolutePath()),
+        testJibExtension.getSkaffold().getSync().getExcludes());
+    Assert.assertEquals(
+        ImmutableSet.of(
+            root.resolve("watch1").toAbsolutePath(), root.resolve("watch2").toAbsolutePath()),
+        testJibExtension.getSkaffold().getWatch().getBuildIncludes());
+    Assert.assertEquals(
+        ImmutableSet.of(root.resolve("watch3").toAbsolutePath()),
+        testJibExtension.getSkaffold().getWatch().getIncludes());
+    Assert.assertEquals(
+        ImmutableSet.of(root.resolve("watch4").toAbsolutePath()),
+        testJibExtension.getSkaffold().getWatch().getExcludes());
   }
 
   @Test