+++ jib-gradle-plugin/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Additionally reads credentials from `~/.docker/.dockerconfigjson` and legacy Docker config (`~/.docker/.dockercfg`). Also searches for `$HOME/.docker/*` (in addition to current `System.get("user.home")/.docker/*`). This may help retrieve credentials, for example, on Kubernetes. ([#2260](https://github.com/GoogleContainerTools/jib/issues/2260))
+- New skaffold configuration options that modify how jib's build config is presented to skaffold ([#2292](https://github.com/GoogleContainerTools/jib/pull/2292)):
+    - `jib.skaffold.watch.buildIncludes`: a list of build files to watch
+    - `jib.skaffold.watch.includes`: a list of project files to watch
+    - `jib.skaffold.watch.excludes`: a list of files to exclude from watching
+    - `jib.skaffold.sync.excludes`: a list of files to exclude from sync'ing
 
 ### Changed